@@ -19,27 +19,27 @@
   "dependencies": {
     "ace-builds": "^1.2.8",
     "async": "2.5.0",
-    "autoprefixer": "7.1.3",
+    "autoprefixer": "7.1.4",
     "bcryptjs": "2.4.3",
-    "benchpressjs": "^1.0.3",
-    "body-parser": "^1.17.2",
+    "benchpressjs": "^1.1.0",
+    "body-parser": "^1.18.2",
     "bootstrap": "^3.3.7",
     "chart.js": "^2.6.0",
     "colors": "^1.1.2",
-    "compression": "^1.7.0",
+    "compression": "^1.7.1",
     "connect-ensure-login": "^0.1.1",
     "connect-flash": "^0.1.1",
     "connect-mongo": "1.3.2",
     "connect-multiparty": "^2.0.0",
-    "connect-redis": "3.3.0",
+    "connect-redis": "3.3.2",
     "cookie-parser": "^1.4.3",
     "cron": "^1.2.1",
-    "cropperjs": "^0.8.1",
+    "cropperjs": "^1.0.0",
     "csurf": "^1.9.0",
     "daemon": "^1.1.0",
-    "express": "^4.15.4",
-    "express-session": "^1.15.5",
-    "express-useragent": "1.0.7",
+    "express": "^4.16.1",
+    "express-session": "^1.15.6",
+    "express-useragent": "1.0.8",
     "html-to-text": "3.3.0",
     "ipaddr.js": "^1.5.2",
     "jimp": "0.2.28",
@@ -49,13 +49,13 @@
     "lodash": "^4.17.4",
     "logrotate-stream": "^0.2.5",
     "lru-cache": "4.1.1",
-    "mime": "^1.4.0",
+    "mime": "^2.0.3",
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
     "mongodb": "2.2.31",
-    "morgan": "^1.8.2",
+    "morgan": "^1.9.0",
     "mousetrap": "^1.6.1",
-    "nconf": "^0.8.4",
+    "nconf": "^0.8.5",
     "nodebb-plugin-composer-default": "6.0.0",
     "nodebb-plugin-dbsearch": "2.0.6",
     "nodebb-plugin-emoji-extended": "1.1.1",
@@ -70,20 +70,20 @@
     "nodebb-theme-slick": "1.1.1",
     "nodebb-theme-vanilla": "7.0.1",
     "nodebb-widget-essentials": "3.0.5",
-    "nodemailer": "4.1.0",
-    "passport": "^0.3.0",
+    "nodemailer": "4.1.1",
+    "passport": "^0.4.0",
     "passport-local": "1.0.0",
-    "postcss": "6.0.10",
-    "postcss-clean": "1.0.3",
+    "postcss": "6.0.12",
+    "postcss-clean": "1.1.0",
     "promise-polyfill": "^6.0.2",
     "prompt": "^1.0.0",
     "redis": "2.8.0",
-    "request": "2.81.0",
-    "rimraf": "2.6.1",
+    "request": "2.83.0",
+    "rimraf": "2.6.2",
     "rss": "^1.2.2",
-    "sanitize-html": "^1.14.0",
+    "sanitize-html": "^1.14.1",
     "semver": "^5.4.1",
-    "serve-favicon": "^2.4.3",
+    "serve-favicon": "^2.4.5",
     "sitemap": "^1.13.0",
     "socket.io": "2.0.3",
     "socket.io-client": "2.0.3",
@@ -92,24 +92,24 @@
     "spdx-license-list": "^3.0.1",
     "string": "^3.3.3",
     "toobusy-js": "^0.5.1",
-    "uglify-js": "^3.0.28",
-    "validator": "8.1.0",
+    "uglify-js": "^3.1.3",
+    "validator": "9.0.0",
     "winston": "^2.3.1",
     "xml": "^1.0.1",
     "xregexp": "3.2.0",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "coveralls": "^2.11.14",
-    "eslint": "^3.12.0",
-    "eslint-config-airbnb-base": "^11.1.0",
-    "eslint-plugin-import": "^2.2.0",
+    "coveralls": "^3.0.0",
+    "eslint": "^4.8.0",
+    "eslint-config-airbnb-base": "^12.0.1",
+    "eslint-plugin-import": "^2.7.0",
     "grunt": "^1.0.1",
     "grunt-contrib-watch": "^1.0.0",
-    "istanbul": "^0.4.2",
-    "jsdom": "^11.2.0",
-    "mocha": "^3.5.0",
-    "mocha-lcov-reporter": "^1.2.0",
+    "istanbul": "^0.4.5",
+    "jsdom": "^11.3.0",
+    "mocha": "^3.5.3",
+    "mocha-lcov-reporter": "^1.3.0",
     "xmlhttprequest": "1.8.0",
     "xmlhttprequest-ssl": "1.5.3"
   },
