@@ -240,7 +240,7 @@ Flags.validate = function (payload, callback) {
 				}
 
 				var minimumReputation = utils.isNumber(meta.config['privileges:flag']) ? parseInt(meta.config['privileges:flag'], 10) : 1;
-					// Check if reporter meets rep threshold (or can edit the target post, in which case threshold does not apply)
+				// Check if reporter meets rep threshold (or can edit the target post, in which case threshold does not apply)
 				if (!editable.flag && parseInt(data.reporter.reputation, 10) < minimumReputation) {
 					return callback(new Error('[[error:not-enough-reputation-to-flag]]'));
 				}
@@ -256,7 +256,7 @@ Flags.validate = function (payload, callback) {
 				}
 
 				var minimumReputation = utils.isNumber(meta.config['privileges:flag']) ? parseInt(meta.config['privileges:flag'], 10) : 1;
-					// Check if reporter meets rep threshold (or can edit the target user, in which case threshold does not apply)
+				// Check if reporter meets rep threshold (or can edit the target user, in which case threshold does not apply)
 				if (!editable && parseInt(data.reporter.reputation, 10) < minimumReputation) {
 					return callback(new Error('[[error:not-enough-reputation-to-flag]]'));
 				}
