@@ -39,8 +39,7 @@ module.exports = function (Plugins) {
 				(Plugins.deprecatedHooks[data.hook] ?
 					'please use `' + Plugins.deprecatedHooks[data.hook] + '` instead.' :
 					'there is no alternative.'
-				)
-			);
+				));
 		} else {
 			// handle hook's startsWith, i.e. action:homepage.get
 			var parts = data.hook.split(':');
@@ -61,7 +60,7 @@ module.exports = function (Plugins) {
 					if (memo && memo[prop]) {
 						return memo[prop];
 					}
-						// Couldn't find method by path, aborting
+					// Couldn't find method by path, aborting
 					return null;
 				}, Plugins.libraries[data.id]);
 
