@@ -46,20 +46,20 @@ describe('admin search', function () {
 	describe('sanitize', function () {
 		it('should strip out scripts', function (done) {
 			assert.equal(
-                search.sanitize('Pellentesque tristique senectus' +
-                    '<script>alert("nope");</script> habitant morbi'),
-                'Pellentesque tristique senectus' +
-                ' habitant morbi'
-            );
+				search.sanitize('Pellentesque tristique senectus' +
+					'<script>alert("nope");</script> habitant morbi'),
+				'Pellentesque tristique senectus' +
+					' habitant morbi'
+			);
 			done();
 		});
 		it('should remove all tags', function (done) {
 			assert.equal(
-                search.sanitize('<p>Pellentesque <b>habitant morbi</b> tristique senectus' +
-                    'Aenean <i>vitae</i> est.Mauris <a href="placerat">eleifend</a> leo.</p>'),
-                'Pellentesque habitant morbi tristique senectus' +
-                'Aenean vitae est.Mauris eleifend leo.'
-            );
+				search.sanitize('<p>Pellentesque <b>habitant morbi</b> tristique senectus' +
+					'Aenean <i>vitae</i> est.Mauris <a href="placerat">eleifend</a> leo.</p>'),
+				'Pellentesque habitant morbi tristique senectus' +
+					'Aenean vitae est.Mauris eleifend leo.'
+			);
 			done();
 		});
 	});
@@ -67,24 +67,20 @@ describe('admin search', function () {
 	describe('simplify', function () {
 		it('should remove all mustaches', function (done) {
 			assert.equal(
-                search.simplify(
-                    'Pellentesque tristique {{senectus}}habitant morbi' +
-                    'liquam tincidunt {mauris.eu}risus'
-                ),
-                'Pellentesque tristique habitant morbi' +
-                'liquam tincidunt risus'
-            );
+				search.simplify('Pellentesque tristique {{senectus}}habitant morbi' +
+					'liquam tincidunt {mauris.eu}risus'),
+				'Pellentesque tristique habitant morbi' +
+					'liquam tincidunt risus'
+			);
 			done();
 		});
 		it('should collapse all whitespace', function (done) {
 			assert.equal(
-                search.simplify(
-                    'Pellentesque tristique   habitant morbi' +
-                    '  \n\n    liquam tincidunt mauris eu risus.'
-                ),
-                'Pellentesque tristique habitant morbi' +
-                '\nliquam tincidunt mauris eu risus.'
-            );
+				search.simplify('Pellentesque tristique   habitant morbi' +
+					'  \n\n    liquam tincidunt mauris eu risus.'),
+				'Pellentesque tristique habitant morbi' +
+					'\nliquam tincidunt mauris eu risus.'
+			);
 			done();
 		});
 	});
